+++ cowork/src/Routes.d.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter } from "react-router-dom";
 import { Locations } from "./pages/locations/locations";
 import { Routes } from "./types/types";
 import { Home } from "./pages/home/home";
-import Login from "./components/login";
+import Login from "./pages/login";
 import { SideBar } from "./components/side-bar/side-bar";
 
 export const router = createBrowserRouter([
@@ -25,6 +25,6 @@ export const router = createBrowserRouter([
   },
   {
     path: Routes.me,
-    element: <div>aaa</div>,
+    element: <div>About me</div>,
   },
 ]);
